Keep failed project creation free of filesystem side effects
Project creation now completes template, feature, and integration preflight before creating a named target directory. Project-name validation lives in a dependency-light shared module so programmatic creation does not import terminal UI code.

Constraint: Shared project creation must remain independent of interactive terminal UI.
Rejected: Keep validation on interactive-wizard.ts | This pulls terminal selection and UI dependencies into programmatic consumers.
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep definition loading and validation before ensureDir so rejected requests do not leave target directories.
Tested: Regression red-green cycle; 6 focused suites (87 steps); format, lint, type checks; dependency graph inspection.
Not-tested: Full repository suite before commit; enforced by the pre-push hook.
