+++ cli/commands/init/interactive-wizard.test.ts
@@ -6,12 +6,7 @@ import "#veryfront/schemas/_test-setup.ts";
 
 import { assertEquals } from "#veryfront/testing/assert.ts";
 import { describe, it } from "#veryfront/testing/bdd.ts";
-import {
-  formatWizardIntro,
-  SETUP_COPY,
-  shouldRunWizard,
-  validateProjectName,
-} from "./interactive-wizard.ts";
+import { formatWizardIntro, SETUP_COPY, shouldRunWizard } from "./interactive-wizard.ts";
 
 describe("interactive-wizard", () => {
   it("starts directly with the setup task", () => {
@@ -23,32 +18,6 @@ describe("interactive-wizard", () => {
     assertEquals(SETUP_COPY.template, "Choose a starter template:");
     assertEquals(SETUP_COPY.runtime, "Select runtime:");
     assertEquals(SETUP_COPY.git, "Initialize Git?");
-  });
-
-  describe("validateProjectName", () => {
-    it("should accept a simple name", () => {
-      assertEquals(validateProjectName("my-app"), null);
-    });
-
-    it("should reject forward slashes", () => {
-      assertEquals(typeof validateProjectName("foo/bar"), "string");
-    });
-
-    it("should reject backslashes", () => {
-      assertEquals(typeof validateProjectName("foo\\bar"), "string");
-    });
-
-    it("should reject dot-dot", () => {
-      assertEquals(typeof validateProjectName(".."), "string");
-    });
-
-    it("should reject single dot", () => {
-      assertEquals(typeof validateProjectName("."), "string");
-    });
-
-    it("should accept dotfiles", () => {
-      assertEquals(validateProjectName(".my-app"), null);
-    });
   });
 
   describe("shouldRunWizard", () => {
